Updated my header and source files
